server/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.7
pgvector==0.2.3
pydantic==2.5.0
openai==1.3.7
sentence-transformers==2.2.2 